fix: resolve 5 CodeQL alerts in icctest V2 code
- cpp/enum-undefined-behavior: cast Severity comparisons to uint8_t
  in icctestParity.cpp and LinuxSandbox.cpp (2 alerts)
- cpp/enum-undefined-behavior: cast icColorSpaceSignature to
  icUInt32Number in IccQualityMetrics.h (1 alert)
- cpp/enum-undefined-behavior: cast Command comparison to uint8_t
  in ArgParser.cpp (1 alert)
- icc/narrow-loop-bound: widen nInput from uint8_t to int in
  ProfileView.cpp CLUT grid overflow check (1 alert)

V1: 813/813 tests pass, V2: all 3 targets build clean.

+++ iccanalyzer-lite/icctest/tools/icctestParity.cpp
@@ -244,7 +244,7 @@ std::optional<Severity> maxFindingSeverity(const CheckResult& result) {
     }
     Severity maxLevel = result.findings.front().level;
     for (const auto& f : result.findings) {
-        if (f.level > maxLevel) {
+        if (static_cast<uint8_t>(f.level) > static_cast<uint8_t>(maxLevel)) {
             maxLevel = f.level;
         }
     }
